Initial implementation of ResourceFile::Save()
Allows writing out of loaded data. This has been tested end-to-end
by loading a .rscr, saving it, and then using binhex to repack it
into a .Hqx and load it on Basilisk II.

To support saving support was added to parse the name table. Currently
system reserved data (240 bytes after the header) is not being preserved
but the application seems to still work fine.