Add MemoryRegion abstraction
Creates a new class which handles safe indexing into a raw pointer.
This allows the logic to be simplified with bounds checking moved
out of the main code path and a templated Copy<> allowing us to
work in terms of types instead of pointers and lengths (sizeof can
be error prone if the wrong struct is passed).

The one downside to this change is that errors are less descriptive
but I think that can be rectified in a future CL. :P