Allow TRY to accept an optional error message
In some cases there is more context at the call site of a TRY
clause than where the Status it consumes was made. By allowing
the call site to provide a more descriptive error message this
context is not lost! :P

This required adding argument overloading for MACROs... support
for which has been added to macro_helpers.h. For more context see
StackOverflow:

https://stackoverflow.com/questions/11761703/overloading-macro-on-number-of-arguments