Add pretty printers for the resource classes
Allows printing all resource groups and their associated resources.
Added a line to main.cc to print the loaded ResourceFile to stdout
and updated logging_internal to ensure LOGs are sent to stderr.

This makes it much easier to see the contents of a Resource Fork

Example:
$ ./rsrcloader ~/BlackBox.rsrc
Group(type: 'CODE'):
  + Resource(id: 1) is 7212 bytes
  + Resource(id: 0) is 24 bytes

Group(type: 'ALRT'):
  + Resource(id: 1000) is 12 bytes
...