Allow specifying a size for MemoryRegion
In preparation of being able to pretty print Resources with their
data this change allows a size to be specified for a given region.

I have chosen to allow a new region to "overflow" its parent's
memory region as long as it does NOT overflow the total size of
memory i.e. sizes are arbitrary except for the Base region's size.

This allows improperly written resource forks to continue to be
read (maybe the sizes specified are a bit wonky?) and only logs
warnings in these situations.

Also keeping track of each MemoryRegion's offet from Base because
it may be useful for pretty printing as well :P