Fix build issues on macOS
absl::StrFormat requires a constexpr for the format string. Fixing
that revealed that const char* STRING_LITERAL = ""; is actually
deprecated... updating these to constexpr char STRING_LITERAL[] = "";
works.

Also had an issue with ResourceTypes being printed jumbled... not
sure how that code was working on Linux (except that I got lucky)
since the char array was not null terminated so std::string was
free to read past the four characters.