Add bin/disasm for disassembling CODE segments
This utility loads CODE segments (currently only segment 1) and
uses Musashi to disassemble the 68K ASM. If an A-Trap is encountered
then instead it will list the name of the trap.

Traps name list was generated with a lot of manual tweaking to give
Trap Op -> Trap Name.

Future work includes:
  - Parsing CODE segment 0 for A5World data [1]
  - Parsing all other CODE segments
  - Parsing CODE segment headers
  - Clean up output and print the raw bytes
  - Add command line flags

[1] http://mirror.informatimago.com/next/developer.apple.com/documentation/mac/runtimehtml/RTArch-116.html#HEADING116-0