Added core/status_main.{cc,h}
absl::Status is used extensively throughout the code base but when
it is finally propagated to the top-level main() we have to convert
it to a LOG(ERROR).

Rather than have that logic duplicated I have moved it into MAIN_LIB
which provides the traditional int main() entry-point implementation
as well as a prototype for a new main entry-point:

  absl::Status Main(const core::Args&);

This allows the TRY/RETURN_IF_ERROR macros to be used directly in
main() with sane handling. It also takes care of calling
absl::ParseCommandLine so that absl::Flags can be used easily (if
needed) and any positional arguments are added to a new core::Args
class which has easy accessors for safely grabbing them.