Add barebones emulation support
This CL is the start of using Musashi to emulate the m68k processor
to run CODE segments for high level emulation of the Mac.

Currently it simply sets up reading from a MemoryRegion and executing
instructions single-step on each user input. This should make it
easier to figure out how to emulate a simple program (single-step
through instructions until we hit an A-Trap to execute natively).