Implement A-Line Exception Handling
Adds support for A-Line exceptions in the emu binary. This the
main mechanism which userspace programs invoke "OS/Toolbox"
functions.

Musashi does not have a simple way to add a callback so instead we
listen for reads to 0x28 which is the address of the A-Line exception
interrupt vector table entry (what a mouth full) [1].

When the read occurs we return a fixed address in high memory which
always contains 0x4E73 (ReTurn from Exception). This takes care of
clearing the Status Register (returning to user-mode from
supervisor-mode [2] used to handle interrupts) and jumping back to
the return address on the stack.

It should be noted that user/supervisor-mode have separate stacks
which both are hidden behind A7 depending on which mode is active [3].

_LoadSeg A-Trap is implemented for near-model segments in this CL
which allows examples/Simple.rsrc to run to completion! :)

examples/Simple.rsrc simply executes SysBeep then ExitToShell.

[1] https://www.eeeguide.com/exceptions-types-of-motorola-68000/
[2] https://www.cs.mcgill.ca/~cs573/fall2002/notes/lec273/lecture21/21_1.htm
[3] http://alanclements.org/68kusersupervisor.html