Generate trap name debug function
Adds a Python generator for creating the GetTrapName() function
(as opposed to checking in a source file). gen/trap_names.txt is
now the source of truth for all traps.

This cuts down on the code in the repo and allows generating the
trap dispatch table in the future. :P