Create a new PopRef stack helper
PopRef<T> pops a pointer to a type T off the stack and then dereferences
it to the type. The previous Pop<absl::string_view> is a special case of
this method and has been renamed PopRef<absl::string_view> as well. This
generalized method will be useful since any none trivial Toolbox
arguments are passed this way on the stack.