Add prototypes for all traps called by examples/Galaxy.rscr
examples/Galaxy.rsrc is small (1748 bytes) demo application that
produces a "star-field" effect which uses 8 Toolbox traps -- this small
size and number of traps makes it an ideal first program to get running.

This CL adds dummy implementations for all of the required traps (simply
popping the correct arguments off the stack and prints them out).

TrapReturn<>() was added in stack_helpers to facilitate the observed
pattern of the caller of traps with a return allocating space on the
stack for the return value before calling the trap -- will need to find
where this is documented and add a reference.