MemoryManager: Add method to allocate a handle for a given region
Utility method AllocateHandleForRegion allocates a "relocatable"
memory block of the correct size and copies data from the provided
MemoryRegion there.

This CL also split up the methods into Allocate (which allocates a
block without a handle), AllocateHandle (which allocates a block with
Allocate() and a Handle for it) and AllocateHandleForRegion. This will
be useful when implementing some of the TRAPs.