TRAP: Dummy System:Get_TrapAddress
Implementing these will require re-designing how traps are implemented
within Cyder so for now they all return a dummy value of 0x3000. It
should be noted that GetTrapAddress is mark as _deprecated_ in more
modern Inside Macintosh as it assumes a single trap dispatch table (as
opposed to the later split dispatch tables for System vs. Toolbox)...