TRAP: Dummy System::SetTrapAddress
Implementing this will require re-designing how traps are implemented
within cyder so only a dummy is provided for now.

It should be noted that SetTrapAddress is mark as _deprecated_ in more
modern Inside Macintosh as it assumes a single trap dispatch table (as
opposed to the later split dispatch tables for System vs. Toolbox)...