Setup TrapManager to allow patching
Previously TrapManager would pop the status register and return address
off the stack as soon as the 1010 (A-Trap) IVT is accessed and directly
execute the trap.

Now that the ability to patch traps is being added trap dispatch has to
become slightly indirect with the 1010 IVT loading an address to the
"handle trap" subroutine onto the stack and jumping to it (with RTS).

This allows a patch to be jumped to instead and then jump to the "handle
trap" dispatch from there. Native subroutines are called from emulated
code by sticking an RTS instruction within emulated memory and watching
for when that 16-bit instruction is being read from memory. When it is
then native code is run and the stack can be manipulated to influence
where the RTS will jump to in the emulated space.

Before executing a trap (native or patch) the stack needs to be updated
to remove the status register and move the return address past the
instruction that triggered it.