TrapManager: TST.W D0 after executing system traps
Most system traps store a return code in D0. Trap Manager is expected to
test the result of D0. After a trap is executed it is common to see a
conditional branch statement based on the results of the TST.W
instruction.

This missing was causing 1000 to ExitToShell as a call to GetHandleSize
was being checked that _some_ size (i.e. > 0) was returned.