gen: Add a tokenizer for typegen
This project requires interacting with types with in the emulated
Mac's memory... This has resulted in lots of repetitive code accessing
each individual field at hardcoded offsets out of memory and manually
converting between big/little endian and helpers to pretty print
values which helps a lot when debugging issues.

This commit is the start of a new DSL to describe structs and types
in memory and autogenerate the previously manually written code to
read/write the structs and create pretty printers.

This is an experiment and may not pan out but could potentially
simplify a lot of the code in cyder (at the cost of writing a compiler
now).