gen: Add code generator to typegen
Adds an initial code generator which writes valid C++ source/header
to define a C++ struct based on the given type definition file and
the debug pretty-printer ostream.

main.py is modified to take a single file and output a file.{cc,h}.
Multiple files may be supported (with includes) in the future but
for now this is good enough.