Migrate to use typegen types everywhere :P
Removes the handwritten emu/system_types.{cc,h} in favor of generated
types from typegen. A polyfill is required to implement ReadType<> for
absl::string_view in order to fully remove stack_helpers.cc as well
(PopRef<> is now a thin layer around ReadType<>).

Python 3.8 is set as the minimum interpreter for CMake.