gen: Add support for array types to typegen
Supports declaring a type [<type>;@<count>] where <type> can be a `type`
or a `struct` and <count> is a field which contains the number of
entries to parse. Currently numeric <count>s are _not_ supported.

Adds a size() method to generated structs which makes parsing possible.

Currently printing to ostreams just displays [<type>;# of entries] i.e.
"[MenuItemResource;3]".