gen: Add support for null-terminated arrays of structs
A struct is considered "null" if the first _field_ is \0. The first
field is the first value read out of memory for a given struct i.e. if
struct A has struct B as its first member then struct A is null if the
first member of struct B is null. This behavior matches patterns used
on Mac OS.