typegen: Pass more information from parser to codegen
Previously `type` being a dict was used to imply that the type is an
array... however this made the code more difficult to understand and
made it difficult to pass information for other types to codegen.

With this change the parser will produce a dict for all types with a
new TypeVariant.{ARRAY,VALUE} used to explicitly identify both cases.