typegen: Remove special syntax for byte width
This was really only useful for copying 3 bytes (1,2,&4 are already
accounted for with the word aligned types) and does not make sense
in many (any?) other situations in the grammar. Instead a `u24` type
is introduced which takes 3 bytes and stores it in a `u32`.

This does not require the special #<bytes> syntax to be added which
complicates the language and we can still handle the packed bytes
case :^)