Remove ResourceMapHeader::type_list_count
`type_list_offset` points to the start of the ResourceTypeList (which
includes the preceding `count` so the same field was being read twice.

It makes more sense for ResourceTypeList to contain this field since it
is technically not guaranteed to appear directly after `name_list_offset`
-- though it happens to in all files used for testing...

This was caught when re-adding support for writing ResourceFiles since
an extra two bytes were being added when this count was written twice --
once as part of ResourceMapHeader and again as part of ResourceTypeList.