typegen: Create separate classes for each parse Expression
The previous implementation was error prone with fields needing to
be referenced with string keys in a dict and lots of enums which
did not make sense in every permutation.

Using @dataclass it is easy to create a bespoke class for each type
with the exact fields that are needed for each.

This also will allow better type-checking with Pylance