typegen: Improve error handling in parser
Previously the parser would advance one token everytime an error
was encountered but that often lead to more errors being logged
because the majority of the time most of the expression was written
correctly with a one thing missing.

This change only advances the token in error conditions when the
start of an expression can not be found in the main loop (but not
in errors within an expression). This seems to create less log spam
when purposefully introducing errors.

Also cleaned up the control flow so that error cases directly raise
ParserException (instead of doing so indirectly through self._error)
and this allowed having the errors collected directly within
Parser::parse() :^)