typegen: Add basic type checking to parser
Added errors if a `type` or `struct` shares a label with an existing
parsed expression or if the type of a `struct` member or `type`
assignment is not defined.

All expressions are parsed in the order that they are defined in the
type definition file so types must be defined in the order they are
used.