typegen: Add "checked" expression types
This CL adds another layer of type expressions into typegen which
will eventually represent what should be codegen'd.

The idea is that codegen is doing too much checking and should
ideally be a very mechanical and straight-forward "rendering" of
an already checked form so that the code can be simpler.

This is the first step towards that goal with "checked" types which
are direct copies of the ones in "parser" in "type_check" and all
references in "codegen" replaced. Later CLs will change the two
sets of types to better suit their new roles.