Fix SDL include to compile on macOS
Performing a clean compile on macOS failed due to #include <SDL2/SDL.h>
not resolving. Checking with sdl2-config:

$ sdl2-config --cflags
-I/usr/local/include/SDL2 -D_THREAD_SAFE

The include directory is relative to SDL2/ and not include/ causing
the error. Fixing the include directories here and will need to check
on Linux when I am back on my workstation.