Add support for HFS Finder flags found in MacBinary II header
These Finder flags are directly copied from the HFS metadata into the
MacBinary header -- this adds an enum definition of them, a function
(ParseFinderFlags()) which takes the packed uint16_t version of the
flags and returns the enum values, and a pretty printer.

All set flags are printed out.

In the future this may be another area of interest to generate with
typegen as the bit-field enums have a lot of boiler plate to be useful.