Fix implementation of "Auto-Pop Stack bit"
The auto-pop bit is set when an intermediate "glue routine" is
called (using a JSR) to invoke an A-Trap. Previously I was clearing
the address added by the JSR but that was not correct... The correct
behavior is to return to where the routine was executed from (or
to exception stack frame IP + 2 to skip over the A-Trap if it is not
set -- the default behavior).

This solution is still a little messy since when the A-Trap is invoked
through the glue routine we lose the pointer to the trap in
PerformTrapDispatch() (previously it could always be gotten from the
return address - 2). The correct solution is to do what a real system
would do and have a unique address for each trap routine. This will
be refactored when I am less tired.