TRAP: Implement save/restore of registers for OS traps
OS traps should be saving D1, D2 and A1 before dispatching and
A0 if the flag indicating that A0 is used as a return is not set.
Only D1 is set by the trap manager (to store the A-Trap for OS
traps so that the flags can be read) but all expected registers
should be saved to account for trap patches. D0 is always a
return register and thus should NOT be saved.

https://dev.os9.ca/techpubs/mac/OSUtilities/OSUtilities-170.html