Add script to regression test cyder emu
This script takes a path to a rsrc(s) and runs them with cyder emu
collecting and sanitizing the log output. This output is then
scraped for logs about executing A-Traps and loading segments.

Comparing the A-Traps and segments present (and their ordering)
against a previous "gold master" version of the logs helps to
detect regressions (and progress being made!).