Store ResourceGroup/Resource as T instead of std::unique_ptr<T>
Previously ResourceGroup and Resource classes were move-only types and
passed around as std::unique_ptr<> to make them compatible with
absl::StatusOr (which requires a copy constructor) but this was fairly
verbose and un-ergonomic.

Since these classes do not actually store the data (only the metadata),
the move-only requirement was unnecessary and an over optimization.

This change also adds a --filter_type argument to rsrcloader exe which
only prints resources with the OSType four-character code provided in
the filter. This change was what ultimately made me realize how
cumbersome the previous code was. :P