TRAP: Fix legacy GetTrapAddress A-Trap implementation
The GetTrapAddress is from when Toolbox and Operating System traps all
resided within one table. As such it can be called with a single byte
trap index which referred to one of 255 entries in the merged table.

Helpfully it is documented [1] which indexes belong to which trap entry
in the split tables so a full A-Trap can be reconstructed to pass to
TrapManager::GetTrapAddress().

[1] http://0.0.0.0:8000/docs/mac/OSUtilities/OSUtilities-185.html