MemoryRegion: Improve debug stream output
The previous implementation of the debug stream always started
printing at the very first value then `bytes_per_line` bytes for
each line but this meant that the output "jumps" around everytime
a new start address occurs (which happens often on the stack). This
change ensures that we always start printing relative to a common
factor (currently using `bytes_per_line`) so that even when values
are added the values from one stack dump to the next are still
aligned relative to each other. Also added a line break at the
8 byte mark. :P