Add boilerplate for a "StackMonitor"
This class will eventually track writes to the stack (and updates to the
stack pointer) so that updates to a particular entity on the stack can
be tracked.

Ideally this class will keep a running log of all of the call-sites that
updated a particular entity and check that reads are aligned to the
values that were originally inserted i.e. reading the back half of one
4-byte entry and front of another should log a WARNING.