Add LogRegionAccess() to setup watches on regions of memory
This method allows more checks to be added dynamically within
CheckReadAccess/CheckWriteAccess for regions of memory. This is useful
while building the emulator to allow partially filled Records to be
returned but ensure that the unfilled Records are not being read or
written to, causing issues down the line.

This is kind of a hacky implementation and should definitely be
improved in the future :P