Include generated files with absolute path (from project root)
Previously the generated files used CMAKE_CURRENT_BINARY_DIR as the
include path to add allowing for generated files to be included without
any path. This lead to some odd errors and hard to diagnose problems
so they are now using CMAKE_BINARY_DIR which allows them to be included
absolutely i.e. "generated_types.h" => "gen/typegen/generated_types.h"
which is easier to understand and more consistent with the other
includes.