Add RESTRICT_FIELD_ACCESS macro and generate field offsets
This macro builds on top of the recently added LogRegionAccess() and
allows restricting access to a whitelisted set of field offsets (which
are now generated by typegen for each type) for a given type stored at
the given address in memory.

This macro allows partially filling Records that are returned to the
emulated program and crashing if any of the fields that have not been
explicitly whitelisted are accessed by the emulation. :)