TRAP: Implement FrontWindow by storing it in GlobalVars::WindowList
`WindowList` stores the top most window. GetNewWindow writes the window
to `WindowList` and Toolbox::FrontWindow reads from `WindowList`.

This assumes that the most recently created window is the top-most
which is a reasonable assumption at this stage in emulation. In the
future the full window list should be maintained (with the top most
window pointing to the next window in z-index order using
WindowRecord::next_window).