TRAP: Add the skeleton of WaitNextEvent
Parses the arguments off of the stack, writes an EventRecord for
`nullEvent`, and returns FALSE. In a future change SDL will be hooked
up to generate events (window update, mouse, keyboard, etc).

Notably Bill does not seem to respect the FALSE return value and always
reads the EventRecord::what field.