Create EventManager to queue window active events on creation
When a window gains focus (transitions to active) then an "updateEvent"
should be returned on a call to Toolbox::WaitNextEvent. This change
creates an EventManager to queue events to be returned and adds an
event to the queue whenever a window is first created (each new window
should get focus).