Refactor MemoryReader out of bin/macbinary.cc
Adding MemoryReader to core allows it to be reused (like in a new PICT
reader :P). The definition of the ReadType template has also been
moved to core (out of typegen) as I think it makes more sense from a
layering perspective and allows ReadType specializations to be written
outside of the context of typegen!