Move endianness swap directly to MemoryRegion
I have resisted having the endianness swap within MemoryRegion for
every read and write since it felt outside the scope of checking for
safe accesses to a region... but all reads and writes require switching
between BIG/little endian for Cyder and not having this has lead to
a number of bugs where an endian swap was missed.

This CL switches all the swaps to take place when reading from memory
and also updates Copy() to Read(). Basically this is just a large
scale clean up of some decisions I made early on in the project that
keep bothering me.