Correct the portBits.bounds and portRect fields
The |portBits.bounds| is meant to link local and global coordinate
systems for QuickDraw by being offset from the global origin so that
the local origin and global origin are both (0, 0). This was causing
issues when drawing windows for 100 Miles which prompted a rereading
of Imaging with QuickDraw -- Figure 2-4 illustrates what is implemented
in this change.