Add BitmapScreen::CopyBits to copy a bitmap image
Adds a method to copy a Rect portion of a bitmap image from an array
to the given Rect within the BitmapScreen. This makes it possible to
render PICT images. :)

Currently only supports images where the source and destination
dimensions are equal (height/width, not x/y). Also made some quality of
life changes to use uint8_t in bitarray_copy() and created more utility
functions for working with Rects.

As a demonstration bin/screen now takes a PICT path as an argument to
render in the lower right hand corner.