Add CHECKs to ensure BitmapScreen::FillRow does not write out of bounds
It is possible to write outside of |bitmap_| which can cause all sorts
of fun crash-y shenanigans... adding CHECKs does not prevent the
crashes but makes them far less mysterious. This also necessitated the
removal of size_t in favor of ints due to a lot of sign-comparison
issues and no real benefit.

bin/screen no longer needs bounds checking of its own during a drag