Store ExitToShell address on stack before launching application
It seems to be expected that there is an address to jump to pushed on
the stack by the Finder (can not find documentation but have observed
it) before launching the application.

When File -> Quit is selected in Bill, Bill jumps to an address saved
from the stack. Mapping this to the address of ExitToShell ensures that
applications have a unified exit path. Added a cute log message to show
that ExitToShell has been called.