TRAP Implement MenuSelect
This trap is meant to have the system take control and update the menu
visuals, only returning the selected item (if any) to the application
on mouse up.

The approach taken here ends the current emulation time slice and pauses
any further execution, turning Cyder into a typical SDL application with
only the event loop and rendering happening. All mouse movement events
are forwarded to the MenuManager which handles drawing the menus.

When the mouse up event is received, MenuManager sets the stack return
value correctly (via a lambda in TrapManager) and resumes emulation
execution from where we left off.

This can certainly be cleaned up but the overall approach feels sound.
Currently the first item in each menu is selected.