BitmapScreen: Allow FillRect to XOr pattern (instead of copying)
This change enables inverting the bitmap when filling based on the
pattern (with each '1' in the pattern inverting the pixel being set
in the bitmap). This makes FillRect() more versatile and enables
the MacOS menu hilite effect to be implemented.

Ran a speed test drawing as many ellipses as possible (with
SDL_RENDER_ACCELERATED) within 5 seconds and on average adding the
lambda and switch selector for mode did not impact performance.