Menu: Move Pop-Up menu logic to its own RAII class
Modeling the pop-up from the MenuBar as an RAII class makes a lot of
sense since it needs to be able to restore state (the desktop pattern
and menu hilite) whenever a different menu is selected or selection
ends.

Now the user can click down on an item in the MenuBar, see a pop-up and
hover over menu item (with proper hilite-ing) and select one! Behavior
mirrors MacOS.