Menu: Fix bug where desktop was inverted under hovered item
Since AutoHiliteRect is an RAII type it will "restore" i.e. invert the
pixels within the rect once it is deconstructed. The |saved_bitmap_|
(generally the desktop pattern) is restored in the deconstructor of
PopUpMenu and _then_ |hovered_rect_| was being deconstructed inverting
that area on the pattern (instead of the menu)...

Resetting the |hovered_rect_| member _before_ restoring |saved_bitmap_|
fixes the issue.