BitmapScreen: Allow CopyBits to copy a portion of the src
Previously |src_rect| represented the dimensions of |src| i.e. it was
a rectangle which covered the entire image in |src|. In order to support
copying a portion of |src| the dimensions need to be passed in
separately.

This allows an arbitrary region of one BitmapScreen to be copied into
another BitmapScreen at a different location :P and thus simplified the
custom logic needed to store a portion of the screen temporarily before
it is overwritten (for pop-up menus and the drag outline).

One issue is that |src_rect| could potentially clip outside of
|src_dims| which would mean there is not enough data to fill |dst_rect|
(since they are assumed to be the same size currently). This is not
causing any issues yet at runtime... I am not sure how this should be
handled.