BitmapImage: Ensure pattern aligns with the left edge of |rect|
In FillRect and FillEllipse the pattern should always start at the top
left corner of the rect that defines them (so that the pattern is
always the same no matter where the rect is located on the screen).

Since the pattern is copied as a byte to a byte-aligned location on the
screen the pattern is not guaranteed to be properly aligned... to fix
this we rotate the pattern right by the number of pixels the rect is off
from being byte aligned.