MemoryManager: Whitelist CurrentA5 and ROM85 globals
CurrentA5 was already being set (stores the initial value of register
A5 a.k.a. the A5 World).

ROM85 determines if we are using an old 64k ROM (from the original 128k
Macintosh) or a newer ROM (from MacAlmanac II):

"Under the old 64K ROM, this was an non-public location that held -1.
Under all newer ROMs, this location always holds a positive value."

It is unclear what specific positive values mean but a value of $3FFF
may imply Color QuickDraw (maybe Mac II?) compatibility (according to
the Executor source)?