TRAP: Implement MoveTo for QuickDraw
Moves the "pen" in the current GrafPort to coordinates given (h, v).

Trying out a fancy new abstraction that scopes a change to a Type in
the emulator by reading it and supplying it to a lambda which then does
some modifications and the updated Type is automatically written out!

Example:

WithPort([](GrafPort& port) {
  // Modify `port` here and any changes will be written on return
  return absl::OkStatus();
});