Fill Window's content region with white
The Retro68k HelloWorld sample CONSOLE window was not drawing a few
pixel border around the content allowing the Desktop grey to be seen in
Cyder. This behavior was not observed within an actual MacOS emulator
so I suspect that the content region is supposed to be cleared to white.

This suspicion is strengthened by the presence of a global var in the
window manager "section" (from MacAlmanac II):

PaintWhite    09DC  word  erase windows before update event?