Fix menu bar being overwritten when a window is moved from under it
The logic to re-draw the grey desktop was changed to account for the
pattern being properly fixed at the origin of a FillRect() instruction
by drawing the full screen _but_ clipping to the bounding rect of the
moved window's structure region...

Unfortunately we did not account for the structure region possibly
being colocated with the menu bar causing the menu bar to be
overwritten. The trick is to "clip" the clip region so that it
respects the bounds of the Desktop.

Maybe TempClipRect should "nest" with each one being set to the
intersection of the previous clip region and the new one?