Track _all_ windows created allowing each to be dragged
Previously only the last created window was checked for if the mouse
was within its drag region. Now we keep a list of all windows and
iterate through each and returning which is selected (if any).

I chose to use WindowRecord::next_window to store the window list as
a linked list (like on original MacOS) because it seemed simpler to
manipulate than using a std::vector (i.e. to reorder the list)...
this may change in the future.

It should be noted that while all windows can now be dragged they do
_not_ redraw correctly yet and not window activate/deactivate events
are generated.