Repaint windows from back-to-front on ordering change
This is quite a hack but works. Ideally we would supply an update
region of the correct shape (to account for overlapping windows, etc)
to the Window during an update event. In the absence of proper Regions
(this will be a later addition), this change uses the Painter's
algorithm [1] to simply invalidate the windows from back-to-front
ordering. When the event is dispatched the entire window is blanked
and frame is redrawn.

This is fairly inefficient but looks good for now :P

[1] https://en.wikipedia.org/wiki/Painter%27s_algorithm