TRAP FrontWindow should use WindowManager as source-of-truth
Writing to WindowList GlobalVar seems unnecessary after giving it some
more thought... While it is documented in Inside Macintosh, it is
ultimately an implementation detail of the OS and should not be directly
relied on (for instance this trap should be called :P).

Since Cyder is meant to be a high-level emulator, it makes more sense
to store implementation details like this within the emulator instead
of trying to keep low-memory globals in sync.