Add support for masking events to EventManager
EventManager allows the application to filter the events returned by
using an event mask. The first, highest-priority event allowed by the
event mask should be returned. This change brings Cyder's EventManager
closer to parity with the expected behavior! :)

Events are now queued into separate queues based on their relative
priority and if there is an event and that event type is allowed by the
event mask it is returned. Events that are filtered out are _not_
removed and may be returned when a different mask is given.