Use C++14 for Cyder moving forward
Finally found a good enough reason to move to a newer version of C++.
typegen structs could not be initialized using initializer lists because
their members are default initialized (integer types are set to 0).

The requirements for allowing initializer list construction was relaxed
in C++14 [0] and while typegen could be modified to generate
constructors for all structs... this is significantly easier and I think
it is about time we start embracing newer language features.

[0] https://blog.petrzemek.net/2015/10/03/structures-and-member-initializers-in-cpp/