Allow TRAP StillDown() to detect if a MouseUp event is in the queue
The StillDown trap is triggered on MouseDown and checks if a MouseUp
is already in the queue (indicating the mouse button is no longer held).

This required some refactors since the NativeBridge received mouse
events directly and does not use the EventQueue so simply queueing an
event when SDL2 gives a MouseUp would result in a random MouseUp in the
queue after a native drag event which is not ideal...

Instead NativeBridge now returns if it has consumed the event and iff
not queues it in the EventManager which is then checked by StillDown()!