Use TrapReturn<bool> in IsDialogEvent()
I missed this call site when I migrated everything over to use the bool
specialization because it was not properly typed... This caused a major
bug because without a type specified it defaulted to writing a uint32_t
instead of a uint16_t like expected.

Using TrapReturn<bool> ensures that the bug is fixed! :P