Swap (x, y) coordinates in SetPt
"1000 Miles" only looks right if the horizontal and vertical position
arguments are swapped around (similar to the issue seen in GetPen).

The arguments _should_ be being popped in the correct order according
to the ABI and function definition, but I suspect there is some
weirdness that alludes me w.r.t. Points?