trap: Implement OpenPort correctly
OpenPort is expected to initialize the given port _and_ set it as
"thePort" -- this is required for Galaxy to work correctly now that
the drawing routines respect the current GrafPort's patterns.

Also set the WindowManager GrafPort as "thePort" in InitWindows since
it _should_ be calling OpenPort... A followup change will refactor
OpenPort into native code to use multiple places.