trap: Implement system time (set GlobalVars::Time & add ReadDateTime)
The system time is stored in the low-memory global "Time" as the number
of seconds since January 1, 1904 (HFS+ Time Epoch). The time is updated
before each emulation timeslice.

ReadDateTime is supposed to fetch the current time from the clock
module on a real machine (returning the seconds since epoch in the
given VAR) and should _not_ be called by an application... but some
software has been found to call it. Since there is no clock chip we
can simply read from the low-memory global! :P