font: Experiment with variable width font rendering
Using the a Python script (create_metrics.py) we can generate a look-up
table of where the glyph starts in each 8x8 bitmap and how wide it is.
These metrics then allow the fixed-width font to be rendered as a
variable-width font. :P

Font rendering looks better with this approach in some case and worse
in others and more parts of the code base will need to be updated for
this to work properly (we assume 8px wide glyphs in a few places).
Disabling this feature by default for now until we have more time to
work on it.

NOTE: One oddity is that with variable-width rendering some glyphs seem
to be mirrored vertically (particularly "i" and "l") and I am unsure why