window: Respect the window variation type
Control whether we draw the title bar based on if it is expected by the
given variation type [1] and only check title bar elements in FindWindow
if one is actually being rendered. :P

This also fixes Window creation by initializing the content, structure,
and update Regions to empty (as opposed to the port_rect) and fill them
using UpdateWindowRegions() for consistency.

[1] http://0.0.0.0:8000/docs/mac/Toolbox/Toolbox-191.html