window: Simplify drawing DragGrayRegion outline
The previous solution was inefficient as it saved the entire screen
under the drag region simply to draw an outline. Now that FrameRect
respects patterns we can simply XOr to draw and undo the outline. :P

This still does not look quite like Mac OS since that one appears to
stay black on the grey background and not alternate between black/white.
This will require some investigation... maybe the pen pattern should
be anchored differently?