Add ReadType<MacBinaryHeader> to read MacBinary II headers
This refactors the MacBinary II reading logic out of bin/macbinary.cc
into a reusable ReadType<MacBinaryHeader> template override. typegen
can not currently be used to generate the ReadType<> because it requires
fixed length Pascal strings and the ability to skip over unused fields.

By writing this more complex type without codegen we are also able to
do some niceties (Finder flags are combined in a single field,
timestamps are converted to absl::Time, Finder position is encoded as
a Point, and values which must be zero can be CHECK'd inline).