Separate creating WindowRecord from storing it
NewWindow implements the TRAP call from Mac OS but when implementing
dialogs it is useful to be able to populate a WindowRecord without
storing it so I split the two functions.

Also fixed an issue in HasTitleBar() helper where kDialog was not
accounted for.

Also added WindowManager::the() because it is a global singleton.

Also implemented WindowManager::ShowWindow() TRAP and used it.