Add initial support for DialogManager
Implements five TRAPs: GetNewDialog, GetDialogItem, SetDialogItem,
                       IsDialogEvent, and ModalDialog

ModalDialog is only partially implemented with many of the item types
not able to be rendered -- userItem in particular requires us to be able
to call into emulated code from native code and return back to native.

This adds enough support to allow all of the dialogs in 1000 Miles and
Bill to at least open (not rendered correctly) and for the button to be
clicked!